+++ packages/amplify-ui-components/src/common/Translations.ts
@@ -109,5 +109,13 @@ export enum AuthStrings {
   SIGN_UP_FAILED = 'Sign Up Failed',
 }
 
-type Translations = AuthErrorStrings | AuthStrings;
-export const Translations = { ...AuthStrings, ...AuthErrorStrings };
+export enum InteractionsStrings {
+  CHATBOT_TITLE = 'ChatBot Lex',
+  TEXT_INPUT_PLACEHOLDER = 'Write a message',
+  VOICE_INPUT_PLACEHOLDER = 'Click mic to speak',
+  CHAT_DISABLED_ERROR = 'Error: Either voice or text must be enabled for the chatbot',
+  NO_BOT_NAME_ERROR = 'Error: Bot name must be provided to ChatBot',
+}
+
+type Translations = AuthErrorStrings | AuthStrings | InteractionsStrings;
+export const Translations = { ...AuthStrings, ...AuthErrorStrings, ...InteractionsStrings };