+++ packages/datastore/src/index.ts
@@ -1,5 +1,3 @@
-import { DataStore, initSchema } from './datastore/datastore';
-import { Predicates } from './predicates';
-
+export { DataStore, DataStoreClass, initSchema } from './datastore/datastore';
+export { Predicates } from './predicates';
 export * from './types';
-export { DataStore, initSchema, Predicates };